New Loading/Title Screen (#4638)
## About The Pull Request

A **significantly** lighter (_and higher-definition_) replacement of the
current loading screen.
Wanted to make something that looked cleaner and cuter.


![loading_screen](https://github.com/user-attachments/assets/083ad5b9-e7e5-48ef-b732-08710cbb3631)

I've not tested it because my .dme won't compile, but - it's a .gif. 

## Why It's Good For The Game

Looks good, quarter of the size of the original, higher fidelity, has a
dono link, nice message.

## Proof Of Testing

N/A

## Changelog
:cl: Bangle
image: added/modified/removed some icons or image
/:cl: