Updates the README to allow NC clause contributions (#4653)
I was going through our assets and noticed that we've accepted some NC
assets over time, and we don't really have an issue with allowing NC
assets since we're not monitizing them.

- Specifies we also accept CY-BY-NC 3.0 Licenses. We've ported sprites
that have a non-commerical clauses (kitty borgs)

The only license we'd have an issue supporting is the No Derivatives
clause.

This isn't a game facing change.