Revert "Replaces the Wizard Dice event with a round-start wizard role" (#4484)
Reverts Bubberstation/Bubberstation#4232

Forgive me for the bad template because one isn't provided.

## Why

I don't think I argued my point well enough because I never suspected
that this PR in question would actually be merged considering the mantra
of "improve, don't remove" that tends to get used around here. That's
legitimately my fault because I didn't want to engage in nitpick mode,
but now that it's merged and the game is worse for it, I now have to
write a novel and look like an ass.

### Part 1: The Die of Fate is not good. Adding it to maint loot is very
bad.
The die of fate can round remove and dust you, or add a permanent
movespeed debuff to you. The table it was added to (one of a kind) is
spawn thrown, meaning that as soon as the dice spawns it is thrown
without warning. If you're extremely unlucky, you can just get dusted.
If you're extremely lucky, you can get wizard, a permant buff that
halves incoming brute/burn damage, a random spellbook, or a copy of the
captain's spare).

Even as a shaft miner hater, I think it belongs in lavaland and
shouldn't be in maint loot. /tg/ does have a die of fate variant that is
in maint loot, but it's a stealthy one and there is no way to know if
it's real except by rolling it.

### Part 2: Wizard just doesn't roll anymore.
In order for wizard to spawn, 35 players need to ready up. On a good
day, at most we see is 25 unless it's an event round. On top of that it
has to roll the very low weight of 2, so it is incredibly unlikely to
get picked with a weight of 2, which ends up making wizard a 4% chance
to be picked. ``2 / (2 + 16 + 8 + 6 + 3 + 8 + 5) = 4%``.

### Part 3: In 60% of cases, people who have wizard enabled as an antag
should probably not be wizard.
The Wizardly Die of Fate was accidentally clever in its design where
people who would not normally enable wizard, or even antag for that
matter, would roll the dice. Even I tried rolling it a few times on days
when I had antag completely disabled because I thought it would be fun.
Talking to people who used the dice have the same experience of "I don't
have antag enabled, but lets go gambling and see what happens!"

Whenever I hear about a shit wizard round, it's played by John
McAntagMain who exchanged an antag token to get a wizard role or someone
who extremely painfully hunted for it and actually ICly harmed other
players just to get at it.

Allowing already established characters to role the dice means that they
are more likely to roleplay as their (usually) roleplay friendly
character instead of (most usually) a throwaway character.

### Part 4: Roundstart wizard is fucking ass.
Back in the day when /tg/ didn't have dynamic, killing the Wizard would
result in a round end, and if the round didn't end, the Wizard had
objectives to hijack the shuttle. Keep in mind that wizards were always
valid, and 9 times out of 10 the crew would go apeshit trying to kill
it.

Compare that to bubberstation, where it is an arp server that has 2.5
hour rounds where it usually takes 30-45 minutes for anyone to
mechanically do anything (research, cargo funds, mining, etc) and maybe
about 20% of the crew are interested in validhunting and the other 80%
are smart enough not to engage the wizard or are picky enough to pretend
it doesn't exist and just do their own thing (there is your 20%/80
statistic). If an early shuttle call is made, admins/players are usually
quick to recall it and send in an ERT instead and prolong everyone's
suffering.

### Part 5: There are better solutions.
The PR that made it an event brought up extremely valid criticisms,
however I strongly believe we can solve these problems by reworking how
the dice works entirely.

If players are hunting for the dice, rework how the dice works and force
it to spawn in areas already populated by players so there is no point
in trying to look for it.

If the event is too frequent, reduce the frequency of the event or even
give the dice a limited amount of roles before it's gone forever.

If there isn't enough risk in rolling the dice, make it so that the dice
is "unfairly" weighted so that you are half as likely to roll a 20 than
other numbers.