[100 HP] Rebalances forging for lower max HP, dagger, axe, sword, imbuing changes (#4504)
## About The Pull Request

This PR introduces multiple changes towards forging weapons. This is
because even currently you can optimize them to crit in 6 blows, which
will become a problem on 100 hp.

The word every is not used to mean literally every single one, as there
will be an exception to almost every rule set here to support an item's
gimmick and it's uniqueness compared to the others, naturally.

## The changes

Every weapon now has **7 force one handed**, with the exception of the
dagger. It has to be kept in mind that perfectly forging them brings
that number up to 10.

Every weapon has a **-5 wound bonus**. Their **inherent 10 AP** helps
bring the amount of hits required down by 1 in the cases of most common
armour types when looking at their damage. Having them be good at
killing, good at wounding and have a unique gimmick of their own would
make them overtuned, especially with imbuing on top of it. We do not
need one melee weapon to rule them all, in short.

**Imbuing** is a powerful tool, and while one can argue that 1 unit per
blow is little, I digress, as after extensively going through chems to
find ones that work best, it's become apparent to me that it's a matter
of game knowledge to go from "Strong" to "Murderous" on the power scale
of a tool through imbuing. For this reason, **imbuing will now take away
10 ap from the weapon**, which serves to make it take 1 more hit crit
thresholds, considering the weapons damage.

A perfectly-forged tool will now have **10 force** when fully complete
with a **2x modifier for wielding them**.


### Making niche picks relevant:
After extensively going through each weapon to see which ones are
actually worth taking and personal experience with which weapons see
use, I have decided to additionally change the 3 weapons left behind
into a state where they are interesting and encourage variety, while
filling a niche of their own among eachother.

**The axe...** 
...had a very, very specific purpose. It was to throw it. It's
undeniably deadly in that aspect, however shields have an increased
chance to deflect thrown items - which while providing some difficulty
to someone like a ninja who can throw these en masse, or a traitor with
a box of 5-6 throwing stars, actually guts the axe which is a
bulky-sized weapon. You cannot carry 5 axes, you have to pray it hits
and then you have succesfully thrown away that weapon.
Due to the mechanics below, it has **had it's embed pain mult halved,
even though it's bulky-sized.**

Not to mention, spears are easier to mass produce, take no time and have
the frankly absurd benefit of having reach.

We are in a **shield meta**, and with the axe being a brute-force
weapon, excuse my lack of historical knowledge here, I've decided to
reinforce that stereotype by making it a fair counter to shields, which
are frankly seen on everything from traitors to security to nukeops.
**The way this works** is that upon swinging at someone who is holding a
shield, if that shield can be broken, 15 additional damage is applied
towards the shield, and if the shield cannot be broken, 15 stamina
damage is applied to the user holding it.

The cases of unbreakable shields are frankly niche, but we should not
exclude it. This does not apply to non-shield blocking items. Breaking
your opponent's weapon would be simply too powerful and make the axe
skyrocket from irrelevant to go-to.

**The dagger...**
... is a tricky one. It currently works as a throwing weapon which as
we've just discussed are not quite as powerful with everyone and their
mother having a blocking implement, and currently it functions as a
lower-damage faster-swingspeed weapon. I digress, this isn't what a
dagger *should* be. With only 8 force at best, the dagger seems measely
and unworth of any attention, however, there is another weapon in the
game which has made itself into the archetypical "assassin's weapon".
Something you stab into people when they don't expect it. The nullblade
is the framework off which the new dagger works, in short, **daggers
deal more damage to prone, stunned, staggered, blind or away-facing
targets. **
They also cause more wounds.
Their attack total is 18 in those circumstances of attacking a
vulnerable target.

**The sword...**
... is outshined by it's brother, the katana. Lacking the raw damage of
the warhammer, the reach of the spear, the Katana's batshit-insane
armour piercing and wounding or even a deadly last-ditch throw, the
sword is mostly a flavour piece.

Due to these factors, the sword was remade into something more fitting
to it's pop-culture identity, similarly to what happened to the other
weapons.

Put short, **it blocks now, with 30% chance when not wielded and 45%
when wielded.** This should hopefully make it a better pick for those
who hope to have a cinematic, parry-filled melee combat experience when
forging the blade of a proper knight.
**The spear...**
... was one of the 3 meta weapons but I'd feel bad if after all of this
it were just the plastitanium spear stats-wise. Yes, even if
holsterable. So it receives **22 throw damage** and has it's wounding
made similar to that of the plastitanium spear.

**The hammer...**
... was tweaked only a little bit. It's total damage when wielded
amounts to 24, the same as chainsaws and fireaxes, with the caveat of
being **25% slower** than other weapons **when hitting mobs**, it hits
just as fast against walls, mechs, etc. Yes it's very simple, even for
my liking, but burst damage isn't healthy for mass-producable weapons
and it's better than having it be JUST more damage.

**The pavise...**
... Was pure bullshit. 75% blockchance with damage reaching the 30's?
The damage was cut in half, up to 15.

## Why It's Good For The Game

Makes a few weapons more interesting than "I swing fast", "I can be
thrown" and "I have more damage" adding depth and player options

Makes weapons comparably powerful so that using "the one that looks
cool" isn't inherently a bad choice

Nerfs the pavise because it's ridiculously strong

## Proof Of Testing


## Changelog

:cl:
add: (Forged) Axes now excel at breaking shields, spears are better for
throwing, daggers are better against vulnerable targets and broadswords
can block.
balance: Imbuing a forged weapon takes away 10 of its AP
balance: Forged weapon damage was lowered across the board, pavise
especially
/:cl: