Re-Buffs the Paddy a bit + Nerfs the Exosuit Disabler (#4263)
## About The Pull Request

TG decided to take the paddy out back and break it's knees after a
CERTAIN INCIDENT. This puts the speed in between the 1 and 2 chassis for
better balance, and ensures we don't have any low pressure lightspeed
tomfoolery anymore.

This also drops the exosuit disabler to 4 bolts with variance instead of
5.
## Why It's Good For The Game

The paddy was unrightfully nerfed into the fucking ground, but at the
same time the disabler was really strong with 5 bolts.

## Proof Of Testing

worked on my machine 

</details>

## Changelog
:cl:
balance: Upped the paddy's speed to sit between the mk1 and mk2 
balance: Nerfed the exosuit disabler to 4 projectiles instead of 5 
/:cl: