Removes the NTBR from the Blueshield (#4600)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
With the Blueshield being encouraged to be less hyper-lethal when
possible, the strongest ballistic gun in the game is probably not great
to have on them.

This PR does two things:
- One, removes the NTBR from the Blueshield by replacing the choice
beacon with the e-gun
- Two, makes the Blueshield start with their special roundstart revolver
because it's cool.
This PR does not remove anything other than the NTBR from the
Blueshield, but does change around where it's stored.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Spawning the Blueshield with their special revolver I think will make
the role feel cooler, without actually changing their firepower
mechanically. It's important to note that the e-gun is in their locker,
and both can be used together.

With the Blueshield switching to a less hyper-lethal role, having the
strongest ballistic weapon in the game tends to hurt that feeling. The
revolver is cooler.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing
Compiles and works
<img width="432" height="585" alt="image"
src="https://github.com/user-attachments/assets/fad06511-eb11-4955-8f9b-9456b1380315"
/>
<img width="386" height="341" alt="image"
src="https://github.com/user-attachments/assets/6a941638-401d-4e22-a1d7-626f14a5a787"
/>
<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->
## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: ReturnToZender
del: The NTBR is no longer available to the Blueshield.
add: The Blueshield now spawns in roundstart with their special
revolver, with their e-gun being in their locker.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->