Adds a borg obedience module (#4670)
## About The Pull Request

Adds a brand new cyborg module to the exofab and lustwish; the borg
obedience module. As a sister module to the dominatrix module, it
functions identically the same, but instead of applying the dominant
aura trait, applies the well trained trait. This also makes it so the
quirk functions on silicons properly.

## Why It's Good For The Game

Some borgs want to have the funny toys but not the dominant aura, or
would prefer to just be able to have access to this trait, even if it
may be a negative one on paper, not to mention it's a bit odd that only
the one side of this two sided quirk is available.

## Proof Of Testing

It compiles, only one dominatrix or obedience module can properly be
used at the same time, borgs react how they should with the quirk.
<details>


![submissable](https://github.com/user-attachments/assets/cee5476b-aa9f-4926-8256-605d33bfa123)

![snapsnap](https://github.com/user-attachments/assets/6a4b416a-d21f-4dff-aabd-d027bfc8292c)

<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl:
add: Obedience cyborg module added, a sister module to the dominatrix
module that applies the well trained quirk
add: Obedience module available in the lustwish and exofab with the
other utility borg modules
code: Updated the dominant aura quirk to work on well trained silicons
/:cl: