Improves disposals code so it actually deals damage to you on direction changes and nerfs disposals damage. (#4562)
## About The Pull Request

Improves disposals code so it actually deals damage to you on direction
changes.
Nerfs disposals damage so that it has a 20% chance to apply instead of
an 80% chance to apply.
You no longer take disposals damage if you're in hard crit or dead.
Disposals damage when it procs deals 4 damage instead of 5.

## Why It's Good For The Game

Making disposals damage work as intended is always good.

With the new 100hp change, people are feeling the (disposals) burn, so
the damage was reduced.

## Proof Of Testing

Draft because I literally coded this from scratch in the github online
code editor.

## Changelog

:cl: BurgerBB
balance: Nerfs disposals damage so that it has a 20% chance to apply
instead of an 80% chance to apply. You no longer take disposals damage
if you're in hard crit or dead. You take 1 less disposals damage when it
applies, now, a total of 4 per turn
fix: Improves disposals code so it actually deals damage to you on
direction changes.
/:cl: