Adds stoats to pet owner quirk! (#4704)
## About The Pull Request
You can now take stoats as pets! Yippee!
## Why It's Good For The Game
Kinda sucks that the only mustelid mob in the game isn't available as
part of the pet owner quirk.

Now they are!
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>
It's literally just adding stoats to the pool of mobs for petowner. It
works as intended.
</details>

## Changelog
:cl:
add: Stoats can now be taken as pets with the Pet Owner Quirk! Enjoy
your little squirmy fella.
/:cl: