Ports the 'modern' player panel from Nova/Splurt (#4586)
## About The Pull Request
Ports https://github.com/NovaSector/NovaSector/pull/5923 which was a
port of https://github.com/SPLURT-Station/S.P.L.U.R.T-tg/pull/9 the new
menu can be accessed via a toggle in the admin preferences in game
options. To see what it looks like, see the attached video in Proof of
Testing

It also adds a new verb, `Toggle-Admin-TGUI` which can be found in the
debug tab and will toggle ***EVERYONE'S*** UI preference to old TGUI it
also reverts the player panel to the old one in case you have the
preference enabled and are unable to access the preferences menu.
## Why It's Good For The Game
It makes things easier on the staff members to find buttons, plus it is
100% optional.
## Proof Of Testing
See the video below.
<details>
<summary>Screenshots/Videos</summary>


https://github.com/user-attachments/assets/4098bc35-95e1-4ae5-91dd-27cb2564c350

</details>

## Changelog
:cl:
admin: Ported a new, toggleable, player panel from NovaSector/SPLURT
admin: Adds a new verb in the debug panel: Toggle Admin TGUI
/:cl:

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>