HOP access to the Blueshield (#4741)
## About The Pull Request

I have granted the Blueshield access to the Head of Personnel's Office,
and NOT to their computer.

## Why It's Good For The Game

The HoP is the second in command, and generally, their office is located
inside the bridge. Thus I think it only makes sense for them to have
access to the office baked in their ID, without having the HoP give them
access every time.

## Proof Of Testing

It works

</details>

## Changelog

:cl:
add: The Blueshield has been granted access to the Head of Personnel’s
office.
/:cl: