Makes Bridge Assistant Permanent instead of a trait (#4537)
## About The Pull Request

Re-adds Bridge Assistant as an available role, for 30 days (Potentially
permanently, if staff approve). This also spawns a coffeemaker aboard
the Bridge each shift. :3

## Why It's Good For The Game

So, per debates on Discord, we've run into concerns around Blueshield
player quality, and concerns that BS is oftentimes very popular because
it's a method of flirting with Command staff. Additionally, BA is very
beloved as a trait, and I'll be honest, it's a relatively harmless role
to add. You solely have access to the Bridge, and you're basically there
to fetch coffee, do paperwork, and be Command's sleeve. You know what I
mean by that. ;3c

My personal opinion is that we should keep it permanently, regardless of
what happens with Blueshield. A role dedicated to literally doing
nothing but idle RP with Command should help for those who want to just
RP and relax, and it's fitting to have an assistant fetching you coffee,
running paperwork for you, and potentially turning traitor. (Right now,
it can't roll traitor, but we can enable that on request.)

## Proof Of Testing

N/A, but it's in the occupation list along with a long list of alt
titles. :3

## Changelog

:cl: Luna/Rykka
add: Adds Bridge Assistant back to Occupations
/:cl:

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>
Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>
Co-authored-by: nevimer <foxmail@protonmail.com>