Fixes two genital sprites (#4748)
## About The Pull Request
Does as the title says, it fixes the aroused sprite of the dentata and
gaping vaginas being missing. I solved this by merging all the layers
that were previously inaccessible, so there may be some oddities.
## Why It's Good For The Game
Fixes a bug.
## Proof Of Testing
It worked on my end.
## Changelog
:cl:
fix: Fixed the aroused state for dentata and gaping vaginas being
invisible.
/:cl: