The penalties for not securing a disk no longer occur if there is no captain (or acting) around because I heard people died to it once. (#4753)
## About The Pull Request

The penalties for not securing a disk no longer occur if there is no
captain (or acting) around.

## Why It's Good For The Game

While securing the disk is a staple of Space Station 13, so is taking a
stapler to the face.

This PR is meant to affect lowpop times when there is not even a captain
assigned to secure the disk, which is pretty common. Players shouldn't
be punished for playing on lowpop when 90% of the population are
probably there to get their 3am gooning session in.

## Proof Of Testing

If it compiles, it works.

## Changelog

:cl: BurgerBB
balance: The penalties for not securing a disk no longer occur if there
is no captain (or acting) around because I heard people died to it once.
/:cl:

---------

Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>