Stunswords cannot be crafted with cargo sabres, renfair claymores and bone swords (#4865)
## About The Pull Request

Upstream PR blacklisted almost all non-true variants, including a lot of
unique ones - this is a very powerful weapon reserved for rare
instances, NOT FOR MASS PRODUCTION.

And with them not having all of the items this PR targets, it slipped by
for us. They (items) shouldn't have been allowed to do this to begin
with.

Scream at me for formatting change on the override.


## Changelog

:cl:
fix: Stunswords cannot be crafted with cargo sabres/bone swords/renfair
swords
/:cl: