Paperwork Implant as Import (#4840)
## About The Pull Request

As title, this PR adds the ability to buy paperwork toolset implants
from the imports tab, alongside the other toolsets.

## Why It's Good For The Game

This is a very silly and gimmicky toolset. It can exclusively be
purchased as an order from medical, in packs of two. It cannot be found
anywhere, or crafted. Wasting a whole departmental order on this toolset
is silly. However, it's pretty fun to use due to the fact that it
includes several tools, like a tape recorder, a clipboard, and a fancy
pen.

Making it so that it can be individually purchased by everyone, just
like the other toolset implants, only seems fair.

## Proof Of Testing

yep
<img width="1019" height="609" alt="immagine"
src="https://github.com/user-attachments/assets/f8a18b8a-c6aa-480e-a8bb-6f4fc7706da1"
/>


## Changelog

:cl:
add: Added paperwork implants as an import from cargo.
/:cl: